Remove out-of-date-test.
